refactor: split containers/dataclasses.py into sub-modules and remove legacy code
- Split monolithic dataclasses.py into individual modules (cell_data, boundaries_data,
  multi_cell_data, shapes_data, image_data, spatial_units_data)
- Update all internal and external imports to point at the new modules
- Delete insitupy/dataclasses/ (legacy package, fully superseded by containers/)
- Delete insitupy/containers/dataclasses.py (backward-compat shim, no remaining dependents)
- Delete insitupy/plotting/spatial_old.py (unreferenced dead code)
- Update 7 test files to import from canonical sub-module paths
