+++ insitupy/containers/__init__.py
@@ -1,5 +1,8 @@
-from .dataclasses import (AnnotationsData, BoundariesData, CellData, ImageData,
-                          MultiCellData, RegionsData, ShapesData,
-                          SpatialUnitsData)
+from .boundaries_data import BoundariesData
+from .cell_data import CellData
+from .image_data import ImageData
+from .multi_cell_data import MultiCellData
+from .shapes_data import AnnotationsData, RegionsData, ShapesData
+from .spatial_units_data import SpatialUnitsData
 from .io import (_read_celldata, _read_multicelldata, _read_shapesdata,
                  read_celldata, read_multicelldata, read_shapesdata)
