--- insitupy/dataclasses/__init__.py
@@ -1,4 +1,0 @@
-from .dataclasses import (AnnotationsData, BoundariesData, CellData, ImageData,
-                          MultiCellData, RegionsData, ShapesData,
-                          SpatialUnitsData)
-from .io import read_celldata, read_multicelldata, read_shapesdata
